Send async tasks to redis with DB transaction
