+++ db/sqlc/store.go
@@ -9,6 +9,7 @@ import (
 type Store interface {
 	Querier
 	TransferTx(ctx context.Context, arg TransferTxParams) (TransferTxResult, error)
+	CreateUserTx(ctx context.Context, arg CreateUserTxParams) (CreateUserTxResult, error)
 }
 
 // Store provides all functions to execute db queries and transaction
@@ -42,102 +43,4 @@ func (store *SQLStore) execTx(ctx context.Context, fn func(*Queries) error) erro
 	}
 
 	return tx.Commit()
-}
-
-// TransferTxParams contains the input parameters of the transfer transaction
-type TransferTxParams struct {
-	FromAccountID int64 `json:"from_account_id"`
-	ToAccountID   int64 `json:"to_account_id"`
-	Amount        int64 `json:"amount"`
-}
-
-// TransferTxResult is the result of the transfer transaction
-type TransferTxResult struct {
-	Transfer    Transfer `json:"transfer"`
-	FromAccount Account  `json:"from_account"`
-	ToAccount   Account  `json:"to_account"`
-	FromEntry   Entry    `json:"from_entry"`
-	ToEntry     Entry    `json:"to_entry"`
-}
-
-// TransferTx performs a money transfer from one account to the other.
-// It creates the transfer, add account entries, and update accounts' balance within a database transaction
-func (store *SQLStore) TransferTx(ctx context.Context, arg TransferTxParams) (TransferTxResult, error) {
-	var result TransferTxResult
-
-	err := store.execTx(ctx, func(q *Queries) error {
-		var err error
-
-		result.Transfer, err = q.CreateTransfer(ctx, CreateTransferParams(arg))
-		if err != nil {
-			return err
-		}
-
-		result.FromEntry, err = q.CreateEntry(ctx, CreateEntryParams{
-			AccountID: arg.FromAccountID,
-			Amount:    -arg.Amount,
-		})
-		if err != nil {
-			return err
-		}
-
-		result.ToEntry, err = q.CreateEntry(ctx, CreateEntryParams{
-			AccountID: arg.ToAccountID,
-			Amount:    arg.Amount,
-		})
-		if err != nil {
-			return err
-		}
-
-		if arg.FromAccountID < arg.ToAccountID {
-			result.FromAccount, result.ToAccount, err = addMoney(
-				ctx,
-				q,
-				arg.FromAccountID,
-				arg.ToAccountID,
-				-arg.Amount,
-				arg.Amount,
-			)
-		} else {
-			result.ToAccount, result.FromAccount, err = addMoney(
-				ctx, q, arg.ToAccountID,
-				arg.FromAccountID,
-				arg.Amount,
-				-arg.Amount,
-			)
-		}
-
-		if err != nil {
-			return err
-		}
-
-		return nil
-	})
-
-	return result, err
-}
-
-// transfer money from account A to account B
-func addMoney(
-	ctx context.Context,
-	q *Queries,
-	accountID1 int64,
-	accountID2 int64,
-	amount1 int64,
-	amount2 int64,
-) (account1 Account, account2 Account, err error) {
-	account1, err = addMoneyAccount(ctx, q, accountID1, amount1)
-	if err != nil {
-		return
-	}
-	account2, err = addMoneyAccount(ctx, q, accountID2, amount2)
-	return
-}
-
-func addMoneyAccount(ctx context.Context, q *Queries, accountID int64, amount int64) (account Account, err error) {
-	account, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
-		ID:     accountID,
-		Amount: amount,
-	})
-	return
 }
