+++ gapi/rpc_create_user.go
@@ -27,14 +27,27 @@ func (server *Server) CreateUser(ctx context.Context, req *pb.CreateUserRequest)
 		return nil, status.Errorf(codes.Internal, "hashing password failed: %v", err)
 	}
 
-	arg := db.CreateUserParams{
-		Username:       req.GetUsername(),
-		HashedPassword: hashedPassword,
-		FullName:       req.GetFullName(),
-		Email:          req.GetEmail(),
+	arg := db.CreateUserTxParams{
+		CreateUserParams: db.CreateUserParams{
+			Username:       req.GetUsername(),
+			HashedPassword: hashedPassword,
+			FullName:       req.GetFullName(),
+			Email:          req.GetEmail(),
+		},
+		AfterCreate: func(user db.User) error {
+			taskPayload := worker.PayloadSendVerifyEmail{
+				Username: user.Username,
+			}
+			taskOpts := []asynq.Option{
+				asynq.ProcessIn(10 * time.Second),
+				asynq.MaxRetry(10),
+				asynq.Queue(worker.QueueCritical),
+			}
+			return server.taskDistributor.DistributeTaskSendVerifyEmail(ctx, taskPayload, taskOpts...)
+		},
 	}
 
-	user, err := server.store.CreateUser(ctx, arg)
+	txResult, err := server.store.CreateUserTx(ctx, arg)
 	if err != nil {
 		if pqErr, ok := err.(*pq.Error); ok {
 			switch pqErr.Code {
@@ -46,21 +59,8 @@ func (server *Server) CreateUser(ctx context.Context, req *pb.CreateUserRequest)
 		return nil, status.Errorf(codes.Internal, "cannot create user: %v", err)
 	}
 
-	taskPayload := worker.PayloadSendVerifyEmail{
-		Username: user.Username,
-	}
-	taskOpts := []asynq.Option{
-		asynq.ProcessIn(10 * time.Second),
-		asynq.MaxRetry(10),
-		asynq.Queue(worker.QueueCritical),
-	}
-	err = server.taskDistributor.DistributeTaskSendVerifyEmail(ctx, taskPayload, taskOpts...)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "faild to distributed task send verify email: %v", err)
-	}
-
 	rsp := &pb.CreateUserResponse{
-		User: convertUserToResponse(user),
+		User: convertUserToResponse(txResult.User),
 	}
 
 	return rsp, nil
